requirements.txt
oauthlib==1.0.3
python-twitter==2.2
requests==2.9.1
requests-oauthlib==0.6.1
wheel==0.29.0
